Rewrite of UI and organization of functions
I rewrote the UI as where it was now, it constantly cleared itself and
replaced things. all CRUD functions now work and the UI is redesigned to
have two views where you can create a record and view an editable table
of the employee database. This project is now entirely MVC!